requirements.txt
Flask
Flask-Cors
httpx
pycryptodomex
protobuf
